+++ packages/snippets/src/ant-design-x/src/components-react/bubble.list.tsx
@@ -3,7 +3,7 @@ import type { AvatarProps } from 'antd';
 import type { RewriteRolesType, RewriteRoleType } from '../types';
 import { Bubble } from '@ant-design/x';
 import { renderVueCompOrSlot } from '@mixte/snippets/ant-design-x/utils';
-import { assertPlainObject } from 'mixte';
+import { assertPlainObject, isFunction } from 'mixte';
 import React from 'react';
 
 interface Props extends Omit<BubbleListProps, 'roles'> {
@@ -29,6 +29,9 @@ export default function (props: Props) {
                 : renderVueCompOrSlot(role.avatar),
               header: renderVueCompOrSlot(role.header),
               footer: renderVueCompOrSlot(role.footer),
+              messageRender: isFunction(role.messageRender)
+                ? (content: string) => renderVueCompOrSlot(role.messageRender!(content))
+                : undefined,
             },
           ];
         }),