+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
   - 🌟 [@mixte/snippets] 为 `Ant Design X` 新增了一个 `defineBubbleListRoles` 工具方法, 用于定义 BubbleList 气泡默认属性 *
+  - 💄 [@mixte/snippets] 使 `Ant Design X` 的 `Bubble` 支持通过 `messageRender` 传递 Vue 组件实现自定义渲染内容 *
 
 ## [v3.4.0-beta.4]
   - 📅 2025-03-20